% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{balancedKm1Folds}
\alias{balancedKm1Folds}
\title{K-1 Balanced Folds}
\usage{
balancedKm1Folds(k, strata, undersample = FALSE, prob = NULL, i = NULL)
}
\arguments{
\item{k}{number of folds}

\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{undersample}{logical, whether to remove elements from the population
in order to try to obtain balanced folds}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}

\item{i}{(optional) integer, fold to be use as holdout data}
}
\value{
A vector containing the indices of the sampled data.
}
\description{
Assigns the population to k balanced folds and returns a random
sample made of elements from k-1 folds.
See the \strong{Details} section below for further information.
}
\details{
Each element in the population is randomly assigned to one of the k
folds so that the percentage of each stratum in the population is balanced
in each fold (see \code{\link{balancedKFolds}} function for further details).
If provided, \code{i} indicates the i-th fold to be considered as holdout data.
If \code{i} is missing, one fold is randomly selected to be the holdout data.
A random sample is then generated by removing the i-th fold and merging the
remaining k - 1 folds together.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define balanced strata
strata = c(rep(1,6),rep(2,6))

#Check ratio
table(strata)/length(strata)

#Assign data to 3 folds
i = balancedKm1Folds(
 strata = strata,
 k = 3
)
#Check indices
i
#Check ratio in the samples made of k-1 folds
table(strata[i])/length(strata[i])

#Define unbalanced strata
strata = c(rep(1,6),rep(2,12))

#Check ratio
table(strata)/length(strata)

#Assign data to 3 folds
i = balancedKm1Folds(
 strata = strata,
 k = 3,
 undersample = TRUE
)
#Check folds
i
#Check ratio in the samples made of k-1 folds
table(strata[i])/length(strata[i])

}
\seealso{
\code{\link{balancedKFolds}}
}
\author{
Alessandro Barberis
}
