% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{balancedSample}
\alias{balancedSample}
\title{Balanced Random Samples}
\usage{
balancedSample(strata, n, replace = FALSE, prob = NULL)
}
\arguments{
\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{n}{positive integer value, the sample size}

\item{replace}{logical, whether to sample with replacement.
Default is \code{FALSE}}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a balanced sample with or without replacement from the
population.
See the \strong{Details} section below for further information.
}
\details{
Balanced sampling is a technique of sampling from a population that
can be partitioned into 'strata' (or 'subpopulations'), where each element in
the population is part of one stratum. It is used to ensure that
subgroups of the population are equally represented in the taken sample.
Different approaches exists to take balanced samples from imbalanced populations.
}
\examples{
\dontrun{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(rep("a", 3),rep("b", 6))

#Check ratio
table(strata)/length(strata)

#Balanced random sample without replacement
i = balancedSample(
  strata = strata,
  n = 6,
  replace = FALSE
)
#Check ratio in the sample
table(strata[i])/length(strata[i])

#Balanced random sample with replacement
i = balancedSample(
  strata = strata,
  n = 8,
  replace = TRUE
)
#Check ratio in the sample
table(strata[i])/length(strata[i])
}
}
\references{
He and Garcia, Learning from Imbalanced Data, IEEE Transactions on Knowledge
and Data Engineering (2009)
}
\author{
Alessandro Barberis
}
\keyword{internal}
