% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{balancedSampleWithoutReplacement}
\alias{balancedSampleWithoutReplacement}
\title{Balanced Random Sample Without Replacement}
\usage{
balancedSampleWithoutReplacement(strata, n, prob = NULL)
}
\arguments{
\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{n}{positive integer value, the sample size}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a balanced sample without replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
When the number of elements per stratum (given by the sample size \code{n}
divided by the number of groups in \code{strata}) is less than the number of
elements in the minority group in \code{strata},
this function implements the so-called "random undersampling", in which
the proportion of the strata in the population is adjusted in the taken sample
by removing elements from the majority stratum, so that each group is balanced.

When the number of elements per stratum is greater than the number of
elements in the minority group in \code{strata}, the function raises an error.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(rep("a", 3),rep("b", 6))

#Check ratio
table(strata)/length(strata)

#Balanced random sample
i = balancedSampleWithoutReplacement(
  strata = strata,
  n = 6
)
#Check indices
i
#Check ratio in the sample
table(strata[i])/length(strata[i])

}
\references{
He and Garcia, Learning from Imbalanced Data, IEEE Transactions on Knowledge
and Data Engineering (2009)
}
\author{
Alessandro Barberis
}
