% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{balancedSampleWithReplacement}
\alias{balancedSampleWithReplacement}
\title{Balanced Random Sample With Replacement}
\usage{
balancedSampleWithReplacement(strata, n, prob = NULL)
}
\arguments{
\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{n}{positive integer value, the sample size}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a balanced sample with replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
This function works when the number of elements per stratum (given by the
sample size \code{n} divided by the number of groups in \code{strata}) is
less/greater than the number of elements in the minority group in
\code{strata}, by taking independent samples with replacement from each group.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(rep("a", 3),rep("b", 6))

#Check ratio
table(strata)/length(strata)

#Balanced random sample with replacement
i = balancedSampleWithReplacement(
  strata = strata,
  n = 8
)
#Check indices
i
#Check ratio in the sample
table(strata[i])/length(strata[i])

}
\references{
He and Garcia, Learning from Imbalanced Data, IEEE Transactions on Knowledge
and Data Engineering (2009)
}
\author{
Alessandro Barberis
}
