% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{bootstrapSample}
\alias{bootstrapSample}
\title{Ordinary Bootstrap Sample}
\usage{
bootstrapSample(N, prob = NULL)
}
\arguments{
\item{N}{positive integer value, the population size}

\item{prob}{(optional) N-length vector of positive numeric values, the
probability weights for obtaining the \code{N} elements}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a bootstrap sample from the population. A bootstrap sample
is a sample taken with replacement having the same size of the population.
See the \strong{Details} section below for further information.
}
\details{
This function is a wrapper to the \code{\link[base]{sample.int}}
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Random sample with replacement
bootstrapSample(N = 10)

}
\author{
Alessandro Barberis
}
