% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{SARC Cardinality Calculation}
\usage{
card(sam)
}
\arguments{
\item{sam}{matrix attained from first-step resampling.}
}
\value{
a list consisting of vectors that contain the information of
        sample partitions "liposarcoma" and "leiomyosarcoma" of the new truth
        sets, including the cardinalities.
}
\description{
The function provides a prerequisite to the second-step resampling.
It calculates the cardinalities of two sample conditions in the new
truth set for TCGA-sarc.
}
\examples{
card(sam)
}
\keyword{cardinality-sarc}
