% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getHoldOutSamples}
\alias{getHoldOutSamples}
\title{Get the left-out data}
\usage{
getHoldOutSamples(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
A list where each item is an integer vector containing the indices of
the elements not sampled from the population.
}
\description{
Get the elements of the population not selected in the samples.
}
\author{
Alessandro Barberis
}
