% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getHoldOutSampleSize}
\alias{getHoldOutSampleSize}
\title{Get the hold-out sample size}
\usage{
getHoldOutSampleSize(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
A vector of integer values, the size of each hold-out sample in \code{x}.
}
\description{
Get the hold-out sample size from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
