% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getNumberOfSamples}
\alias{getNumberOfSamples}
\title{Get the number of taken samples}
\usage{
getNumberOfSamples(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
The number of taken samples.
}
\description{
Get the number of taken samples from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
