% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getPopulationSize}
\alias{getPopulationSize}
\title{Get the population size}
\usage{
getPopulationSize(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
An integer value, the population size.
}
\description{
Get the population size from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
