% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getRemovedElements}
\alias{getRemovedElements}
\title{Get the removed elements}
\usage{
getRemovedElements(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
A vector of integer values, the indices of elements removed from the
population.
}
\description{
Get the elements removed from the population before sampling
from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
