% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{getResamplingFunction}
\alias{getResamplingFunction}
\title{Get Resampling Function}
\usage{
getResamplingFunction(
  id = c("rswor", "srswor", "stratified_rswor", "balanced_rswor", "permutation",
    "kfolds", "stratified_kfolds", "balanced_kfolds", "leave_p_out", "leave_one_out",
    "rswr", "srswr", "stratified_rswr", "balanced_rswr", "bootstrap")
)
}
\arguments{
\item{id}{character string, one of the supported resampling techniques}
}
\value{
A resampling function:

\describe{
\item{\code{"rswor"               }}{\code{\link{repeatedSampleWithoutReplacement}}}
\item{\code{"srswor"              }}{\code{\link{repeatedSimpleRandomSampleWithoutReplacement}}}
\item{\code{"stratified_rswor"    }}{\code{\link{repeatedStratifiedSampleWithoutReplacement}}}
\item{\code{"balanced_rswor"      }}{\code{\link{repeatedBalancedSampleWithoutReplacement}}}
\item{\code{"permutation"         }}{\code{\link{repeatedPermutationSample}}}
\item{\code{"kfolds"              }}{\code{\link{repeatedRandomKm1Folds}}}
\item{\code{"stratified_kfolds"   }}{\code{\link{repeatedStratifiedKm1Folds}}}
\item{\code{"balanced_kfolds"     }}{\code{\link{repeatedBalancedKm1Folds}}}
\item{\code{"leave_p_out"         }}{\code{\link{repeatedLeavePOut}}}
\item{\code{"leave_one_out"       }}{\code{\link{repeatedLeaveOneOut}}}
\item{\code{"rswr"                }}{\code{\link{repeatedSampleWithReplacement}}}
\item{\code{"srswr"               }}{\code{\link{repeatedSimpleRandomSampleWithReplacement}}}
\item{\code{"stratified_rswr"     }}{\code{\link{repeatedStratifiedSampleWithReplacement}}}
\item{\code{"balanced_rswr"       }}{\code{\link{repeatedBalancedSampleWithReplacement}}}
\item{\code{"bootstrap"           }}{\code{\link{repeatedBootstrapSample}}}
}
}
\description{
This function is a dispatcher for the resampling function selected
in input.
}
\examples{
f = getResamplingFunction(id = 'rswor')

}
\author{
Alessandro Barberis
}
