% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getSamples}
\alias{getSamples}
\title{Get the sampled data}
\usage{
getSamples(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
A list of samples repeatedly taken from the population. Each
item of the list is an integer vector containing the indices of the
elements sampled from the population.
}
\description{
Get the list of samples from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
