% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getSampleSize}
\alias{getSampleSize}
\title{Get the sample size}
\usage{
getSampleSize(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
A vector of integer values, the size of each sample in \code{x}.
}
\description{
Get the sample size from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
