% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{getSamplingFunction}
\alias{getSamplingFunction}
\title{Get Sampling Function}
\usage{
getSamplingFunction(
  id = c("rswor", "srswor", "stratified_rswor", "balanced_rswor", "permutation",
    "kfolds", "stratified_kfolds", "balanced_kfolds", "leave_p_out", "leave_one_out",
    "rswr", "srswr", "stratified_rswr", "balanced_rswr", "bootstrap")
)
}
\arguments{
\item{id}{character string, one of the supported sampling techniques}
}
\value{
A sampling function:

\describe{
\item{\code{"rswor"               }}{\code{\link{sampleWithoutReplacement}}}
\item{\code{"srswor"              }}{\code{\link{simpleRandomSampleWithoutReplacement}}}
\item{\code{"stratified_rswor"    }}{\code{\link{stratifiedSampleWithoutReplacement}}}
\item{\code{"balanced_rswor"      }}{\code{\link{balancedSampleWithoutReplacement}}}
\item{\code{"permutation"         }}{\code{\link{permutationSample}}}
\item{\code{"kfolds"              }}{\code{\link{randomKm1Folds}}}
\item{\code{"stratified_kfolds"   }}{\code{\link{stratifiedKm1Folds}}}
\item{\code{"balanced_kfolds"     }}{\code{\link{balancedKm1Folds}}}
\item{\code{"leave_p_out"         }}{\code{\link{leavePOutSample}}}
\item{\code{"leave_one_out"       }}{\code{\link{leaveOneOutSample}}}
\item{\code{"rswr"                }}{\code{\link{sampleWithReplacement}}}
\item{\code{"srswr"               }}{\code{\link{simpleRandomSampleWithReplacement}}}
\item{\code{"stratified_rswr"     }}{\code{\link{stratifiedSampleWithReplacement}}}
\item{\code{"balanced_rswr"       }}{\code{\link{balancedSampleWithReplacement}}}
\item{\code{"bootstrap"           }}{\code{\link{bootstrapSample}}}
}
}
\description{
This function is a dispatcher for the sampling function selected
in input.
}
\examples{
f = getSamplingFunction(id = 'rswor')

}
