% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{getSamplingMethodId}
\alias{getSamplingMethodId}
\title{Get the sampling method id}
\usage{
getSamplingMethodId(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
A character string, the sampling method id.
}
\description{
Get the sampling method id from a \code{\link{resampling}} object.
}
\author{
Alessandro Barberis
}
