% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{is.resampling}
\alias{is.resampling}
\title{Check a \code{resampling} object}
\usage{
is.resampling(x)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}
}
\value{
Returns \code{TRUE} if its argument is a valid \code{\link{resampling}} object.
}
\description{
Function to check if an object is a \code{\link{resampling}} object.
}
\examples{
#valid object
is.resampling(x = resampling())

#invalid object
is.resampling(x = 3)

}
\author{
Alessandro Barberis
}
