% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{isValidCvInput}
\alias{isValidCvInput}
\title{K-Folds Input Validator}
\usage{
isValidCvInput(N, k, strata, undersample, prob)
}
\arguments{
\item{N}{population size}

\item{k}{number of folds}

\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{undersample}{logical, whether to remove elements from the population
in order to try to obtain balanced folds}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}
}
\value{
\code{TRUE} if the arguments are valid, or a string
containing the errors.
}
\description{
This function checks the input provided to the k-folds sampling
functions.
}
\examples{
\dontrun{
#Valid input
isValidCvInput(
 strata=c(1,1,1,2,2,2,2,2,2),
 k=3
)

#Invalid input
isValidCvInput(
 N=2,
 k=1
)

isValidCvInput(
 k=1
)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
