% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{isValidLpOSampleInput}
\alias{isValidLpOSampleInput}
\title{Leave-p-out Sample Input Validator}
\usage{
isValidLpOSampleInput(N, p, prob)
}
\arguments{
\item{N}{positive integer value, the population size}

\item{p}{integer, the number of elements to holdout}

\item{prob}{(optional) N-length vector of positive numeric values, the
probability weights for obtaining the \code{N} elements}
}
\value{
\code{TRUE} if the arguments are valid, or a string
containing the errors.
}
\description{
This function checks the input provided to the leave-p-out sampling
functions.
}
\examples{
\dontrun{

#Valid input
isValidLpOSampleInput(
 N=10,
 p=3
)

#Invalid input
isValidLpOSampleInput(
 N=10,
 p=10
)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
