% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{isValidSampleInput}
\alias{isValidSampleInput}
\title{Sample Input Validator}
\usage{
isValidSampleInput(N, n, prob, replace)
}
\arguments{
\item{n}{positive integer value, the sample size}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}

\item{replace}{logical, whether to sample with replacement.
Default is \code{FALSE}}
}
\value{
\code{TRUE} if the arguments are valid, or a string
containing the errors.
}
\description{
This function checks the input provided to the random sampling
functions.
}
\examples{
\dontrun{
#Valid input
isValidSampleInput(
 N=10,
 n=3
)

#Invalid input
isValidSampleInput(
 N=10,
 n=11,
 replace = FALSE
)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
