% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{isValidStratifiedSampleInput}
\alias{isValidStratifiedSampleInput}
\title{Sample Input Validator}
\usage{
isValidStratifiedSampleInput(strata, n, prob, replace)
}
\arguments{
\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{n}{positive integer value, the sample size}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}

\item{replace}{logical, whether to sample with replacement.
Default is \code{FALSE}}
}
\value{
\code{TRUE} if the arguments are valid, or a string
containing the errors.
}
\description{
This function checks the input provided to the random sampling
functions.
}
\examples{
\dontrun{
#Valid input
isValidStratifiedSampleInput(
 strata=c(1,1,1,2,2,2,2,2,2),
 n=3
)

#Invalid input
isValidStratifiedSampleInput(
 n=11,
 replace = FALSE
)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
