% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{km1Folds}
\alias{km1Folds}
\title{K-1 folds}
\usage{
km1Folds(foldid, i = NULL)
}
\arguments{
\item{foldid}{vector of k fold ids}

\item{i}{(optional) integer, fold to be use as holdout data}
}
\value{
A vector containing the indices of the sampled data.
}
\description{
Takes a population assigned to k folds and returns a random sample
made of elements from k-1 folds.
}
\details{
The returned sample is a vector of indices generated by removing the
holdout fold and merging the remaining k - 1 folds together.
If \code{i} is missing, one fold is randomly selected to be the holdout data.
}
\examples{
\dontrun{
#set seed for reproducibility
set.seed(seed = 5381L)

#Assign data to 3 folds
km1Folds(
 foldid = randomKFolds(N = 10, k = 3)
)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
