% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{leaveOneOutSample}
\alias{leaveOneOutSample}
\title{Leave-One-Out Sampling}
\usage{
leaveOneOutSample(N)
}
\arguments{
\item{N}{positive integer value, the population size}
}
\value{
A vector of length \code{N-1} containing the index of the
computed random set of observations.
}
\description{
Takes a random sample of size \code{N-1} from the population.
See the \strong{Details} section below for further information.
}
\details{
The leave-one-out is a particular case of the leave-p-out, where
\code{p = 1}.
A random sample of size \code{1} is taken from the population and used as holdout
data.
This function returns a sample of size \code{N-1} obtained by removing the holdout
sample from the population.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Take one sample leaving out 1 element
leaveOneOutSample(N = 10)

}
\seealso{
\code{\link{leavePOutSample}}
}
\author{
Alessandro Barberis
}
