% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSamplingMethodIds}
\alias{listAvailableSamplingMethodIds}
\title{Available Random Sampling Method IDs}
\usage{
listAvailableSamplingMethodIds(x = c("all", "rswor", "rswr"))
}
\arguments{
\item{x}{character, whether to return all sampling methods (\code{all}) or only the
methods using sampling with (\code{rswr}) or without (\code{rswor}) replacement}
}
\value{
A vector containing the sampling method identifiers.
}
\description{
This function returns the currently available
random sampling methods.
}
\author{
Alessandro Barberis
}
\keyword{internal}
