% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSamplingWithoutReplacementMethodIds}
\alias{listAvailableSamplingWithoutReplacementMethodIds}
\title{Available Random Sampling Without Replacement Method Ids}
\usage{
listAvailableSamplingWithoutReplacementMethodIds()
}
\value{
A vector containing the sampling method identifiers.
}
\description{
This function returns the ids of the currently available methods
for random sampling without replacement.
}
\author{
Alessandro Barberis
}
\keyword{internal}
