% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSamplingWithReplacementMethodIds}
\alias{listAvailableSamplingWithReplacementMethodIds}
\title{Available Random Sampling With Replacement Method Ids}
\usage{
listAvailableSamplingWithReplacementMethodIds()
}
\value{
A vector containing the sampling method identifiers.
}
\description{
This function returns the ids of the currently available methods
for random sampling with replacement.
}
\author{
Alessandro Barberis
}
\keyword{internal}
