% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listSamplingFunctionNames}
\alias{listSamplingFunctionNames}
\title{Available Random Sampling Functions}
\usage{
listSamplingFunctionNames(x = c("all", "rswor", "rswr"))
}
\arguments{
\item{x}{character, whether to return all sampling methods (\code{all}) or only the
methods using sampling with (\code{rswr}) or without (\code{rswor}) replacement}
}
\value{
A data frame with two columns:

\describe{
\item{id}{the id of the random sampling method, to be used in the function calls}
\item{name}{the name of the random sampling function}
}
}
\description{
This function returns the currently available
random sampling functions.
}
\examples{

listSamplingFunctionNames()

}
\author{
Alessandro Barberis
}
