% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{mapStrataToColorsWithCirclize}
\alias{mapStrataToColorsWithCirclize}
\title{Map Strata to Colors}
\usage{
mapStrataToColorsWithCirclize(strata, ...)
}
\arguments{
\item{strata}{vector of stratification variables}

\item{...}{further arguments to \code{\link[circlize:rand_color]{rand_color}}}
}
\value{
A vector of colors.
}
\description{
This function maps the strata variables to colors.
}
\author{
Alessandro Barberis
}
\keyword{internal}
