% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{mapStrataToColorsWithColorBrewer}
\alias{mapStrataToColorsWithColorBrewer}
\title{Map Strata to Colors}
\usage{
mapStrataToColorsWithColorBrewer(
  strata,
  palette = c("auto", "Blues", "BuGn", "BuPu", "GnBu", "Greens", "Greys", "Oranges",
    "OrRd", "PuBu", "PuBuGn", "PuRd", "Purples", "RdPu", "Reds", "YlGn", "YlGnBu",
    "YlOrBr", "YlOrRd", "Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1",
    "Set2", "Set3", "BrBG", "PiYG", "PRGn", "PuOr", "RdBu", "RdGy", "RdYlBu", "RdYlGn",
    "Spectral")
)
}
\arguments{
\item{strata}{vector of stratification variables}

\item{palette}{character, a RColorBrewer palette name. If \code{auto}, an automatic
selection based on the number of stratification variables is made}
}
\value{
A vector of colors.
}
\description{
This function maps the strata variables to colors.
}
\author{
Alessandro Barberis
}
\keyword{internal}
