% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{permutationSample}
\alias{permutationSample}
\title{Permutation Samples}
\usage{
permutationSample(N)
}
\arguments{
\item{N}{positive integer value, the population size}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a permutation sample.
See the \strong{Details} section below for further information.
}
\details{
This function is a wrapper to the \code{\link[base]{sample.int}}
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Random sequence of N elements from 1 to N
permutationSample(N = 10)

}
\author{
Alessandro Barberis
}
