% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{plot.resampling}
\alias{plot.resampling}
\title{Plot a \code{resampling} object}
\usage{
\method{plot}{resampling}(
  x,
  ...,
  cell.width = grid::unit(10, "mm"),
  cell.height = cell.width,
  col = c("purple", "lightblue", "gray"),
  border = NA,
  rect_gp = grid::gpar(col = "white"),
  heatmap.legend.title = "type",
  strata = NULL,
  strata.col = NULL
)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}

\item{...}{additional plot arguments}

\item{cell.width, cell.height}{width and height of the cells in the heatmap}

\item{col}{length-3 vector containing the colors for the heatmap}

\item{border}{Whether draw border. The value can be logical or a string of color.}

\item{rect_gp}{Graphic parameters for drawing rectangles (for heatmap body). The value should be specified by \code{\link[grid]{gpar}} and \code{fill} parameter is ignored.}

\item{heatmap.legend.title}{character string, title of the heatmap legend}

\item{strata}{(optional) vector of stratification variables}

\item{strata.col}{(optional) vector of colors for the stratification variables}
}
\value{
A \code{\link[ComplexHeatmap:Heatmap-class]{Heatmap-class}} object.
}
\description{
Plot a \code{\link{resampling}} object.
See the \strong{Details} section below for further information.
}
\details{
This function plot the samples taken from the population as an heatmap
where each row corresponds to a different sample, while the columns represent
the elements in the population.
The elements belong to one of the following groups:
\itemize{
\item \code{sample}: the sample taken from the population
\item \code{holdout}: the data not selected in the sample
\item \code{removed}: the data removed from the population and not considered for sampling
}

If the sampling technique allows for replacement, the number of repetition are
shown in the cells of the heatmap.

If \code{strata} is provided, an annotation is added to the heatmap.
}
\examples{
#check needed package
require(ComplexHeatmap, quietly = TRUE)

#resampling object
x = resampling(
 method = 'rswor',
 N = 10,
 samples = list(c(1,5,7), c(6,2,9))
)

#plot
plot(x)

#plot with strata
plot(x, strata = c(rep("a", 5), rep("b", 5)))

}
