% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{+.resampling}
\alias{+.resampling}
\title{Combine samples from \code{resampling} objects}
\usage{
\method{+}{resampling}(e1, e2)
}
\arguments{
\item{e1, e2}{object of class \code{\link{resampling}}}
}
\value{
An object of class \code{\link{resampling}} where the \code{samples} slot is the
combination of the samples of \code{e1} and \code{e2}.
}
\description{
Combine the list of samples from \code{\link{resampling}} objects obtained
using the same sampling method and population size.
}
\author{
Alessandro Barberis
}
\keyword{internal}
