% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{print.resampling}
\alias{print.resampling}
\title{Print a \code{resampling} object}
\usage{
\method{print}{resampling}(x, nrows = 5L, nchars = 10L, ...)
}
\arguments{
\item{x}{an object of class \code{\link{resampling}}}

\item{nrows}{integer, the number of samples to show. If more than one value is
provided, it indicates the indices of samples to show}

\item{nchars}{integer, the maximum number of characters to show in the summary
table}

\item{...}{additional print arguments}
}
\value{
Silently return \code{x}.
}
\description{
Print a summary of the \code{\link{resampling}} object.
}
\examples{
#resampling object
x = resampling(
 method = 'rswor',
 N = 10,
 samples = list(c(1,5,7), c(6,2,9))
)

#print
print(x)

}
