% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{randomKFolds}
\alias{randomKFolds}
\title{Random K Folds}
\usage{
randomKFolds(N, k)
}
\arguments{
\item{N}{population size}

\item{k}{number of folds}
}
\value{
A vector of length \code{N} containing the fold ids.
}
\description{
Randomly assigns the population to k folds.
See the \strong{Details} section below for further information.
}
\details{
Each element in the population is randomly assigned to one of the
k folds. Internally, it uses the \code{\link[base]{sample}} function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Assign data to 3 folds
randomKFolds(N = 10, k = 3)

}
\author{
Alessandro Barberis
}
