% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{randomKm1Folds}
\alias{randomKm1Folds}
\title{K-1 Random Folds}
\usage{
randomKm1Folds(k, N, i = NULL)
}
\arguments{
\item{k}{number of folds}

\item{N}{population size}

\item{i}{(optional) integer, fold to be use as holdout data}
}
\value{
A list of length \code{k} where each element is a vector
containing the indices of the sampled data.
}
\description{
Randomly assigns the population to k folds and returns a random sample
made of elements from k-1 folds.
See the \strong{Details} section below for further information.
}
\details{
Each element in the population is randomly assigned to one of the k
folds by using the \code{\link{randomKFolds}} function.
If provided, \code{i} indicates the i-th fold to be considered as holdout data.
If \code{i} is missing, one fold is randomly selected to be the holdout data.
A random sample is then generated by removing the i-th fold and merging the
remaining k - 1 folds together.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Assign data to 3 folds
randomKm1Folds(N = 10, k = 3)

}
\seealso{
\code{\link{randomKFolds}}
}
\author{
Alessandro Barberis
}
