% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{randomSample}
\alias{randomSample}
\title{Simple Random Samples}
\usage{
randomSample(N, n, replace = FALSE, prob = NULL)
}
\arguments{
\item{N}{positive integer value, the population size}

\item{n}{positive integer value, the sample size}

\item{replace}{logical, whether to sample with replacement.
Default is \code{FALSE}}

\item{prob}{(optional) N-length vector of positive numeric values, the
probability weights for obtaining the \code{N} elements}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a simple random sample from the population.
See the \strong{Details} section below for further information.
}
\details{
This function is a wrapper to the \code{\link[base]{sample.int}}
function.
}
\examples{
\dontrun{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Simple random sample without replacement
randomSample(N = 10, n = 5, replace = FALSE)

#Simple random sample with replacement
randomSample(N = 10, n = 5, replace = TRUE)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
