% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang}
\alias{rang}
\title{First-step Resampling Function}
\usage{
rang(SEED, N, M, n)
}
\arguments{
\item{SEED}{a single value, interpreted as an integer.}

\item{N}{a positive integer giving the number of iterations.}

\item{M}{a positive integer giving the range for resampling.}

\item{n}{a non-negative integer giving the number of items to choose.}
}
\value{
a matrix that each row represents the indices of samples for each iteration.
}
\description{
This function allows you to randomly generate a desired number
of samples for certain times from your data sets.
}
\examples{
rang(1, 10, 100, 5)
}
\keyword{resampling}
