% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedBootstrapSample}
\alias{repeatedBootstrapSample}
\title{Repeated Bootstrap Samples}
\usage{
repeatedBootstrapSample(k, N, prob = NULL)
}
\arguments{
\item{k}{integer, the number of repeated samples to generate}

\item{N}{positive integer value, the population size}

\item{prob}{(optional) N-length vector of positive numeric values, the
probability weights for obtaining the \code{N} elements}
}
\value{
A list of length \code{k} where each element
is a vector of indices representing the sampled data.
}
\description{
Takes repeated samples with replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
This function takes repeated samples by using the \code{\link{replicate}}
function and the \code{\link{bootstrapSample}} sampling
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Take two samples with replacement
repeatedBootstrapSample(
 k = 2,
 N = 10
)

}
\seealso{
\code{\link{bootstrapSample}}
}
\author{
Alessandro Barberis
}
