% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedKm1Folds}
\alias{repeatedKm1Folds}
\title{Repeated k-1 folds}
\usage{
repeatedKm1Folds(foldid)
}
\arguments{
\item{foldid}{vector of k fold ids}
}
\value{
A list of length \code{k} where each element is a vector
containing the indices of the sampled data.
}
\description{
Takes a population assigned to k folds and return a list of k
unique samples made of elements from k-1 folds.
}
\details{
A list of length k is created from the folds, so that the i-th
item of the list is a vector of indices generated by removing the i-th fold
and merging the remaining k - 1 folds together.
}
\examples{
\dontrun{
#set seed for reproducibility
set.seed(seed = 5381L)

#Assign data to 3 folds
repeatedKm1Folds(
 foldid = randomKFolds(N = 10, k = 3)
)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
