% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedLeaveOneOut}
\alias{repeatedLeaveOneOut}
\title{Repeated Leave-One-Out}
\usage{
repeatedLeaveOneOut(N)
}
\arguments{
\item{N}{positive integer value, the population size}
}
\value{
A list of length \code{N} where each element is a vector
containing the indices of the sampled data.
}
\description{
Takes repeated samples from the population.
See the \strong{Details} section below for further information.
}
\details{
The leave-one-out is a particular case of the leave-p-out, where
\code{p = 1}.
Samples are repeatedly taken from the population until each element in the
population is considered as the holdout data.
This function returns a list of \code{N} samples of size \code{N-1} obtained by removing
the holdout samples from the population.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Assign data to N folds
repeatedLeaveOneOut(N = 10)

}
\seealso{
\code{\link{repeatedLeavePOut}}
}
\author{
Alessandro Barberis
}
