% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedLeavePOut}
\alias{repeatedLeavePOut}
\title{Repeated Leave-P-Out}
\usage{
repeatedLeavePOut(N, p)
}
\arguments{
\item{N}{positive integer value, the population size}

\item{p}{integer, the number of elements to holdout}
}
\value{
A list of length \eqn{\binom{N}{p}} where each element is a vector
containing the indices of the sampled data.
}
\description{
Takes repeated samples from the population.
See the \strong{Details} section below for further information.
}
\details{
Samples of size \code{p} are repeatedly taken from the population until
all the possible combinations of \code{p} elements are considered. These samples
are used as holdout data. This function returns a list of
\eqn{\binom{N}{p} = \frac{N!}{p!(N-p)!}}
samples of size \code{N-p} obtained by removing the holdout samples from the
population.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Repeatedly sample leaving out p elements each time
repeatedLeavePOut(N = 5, p = 2)

#Equivalent to leave-one-out
repeatedLeavePOut(N = 5, p = 1)

}
\author{
Alessandro Barberis
}
