% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedPermutationSample}
\alias{repeatedPermutationSample}
\title{Repeated Permutation Samples}
\usage{
repeatedPermutationSample(k, N)
}
\arguments{
\item{k}{integer, the number of repeated samples to generate}

\item{N}{positive integer value, the population size}
}
\value{
A list of length \code{k} where each element
is a vector of indices representing the sampled data.
}
\description{
Takes repeated samples without replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
This function takes repeated samples by using the \code{\link{replicate}}
function and the \code{\link{permutationSample}} sampling
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Take two samples without replacement
repeatedPermutationSample(
 k = 2,
 N = 10
)

}
\seealso{
\code{\link{permutationSample}}
}
\author{
Alessandro Barberis
}
