% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedRandomKm1Folds}
\alias{repeatedRandomKm1Folds}
\title{Repeated k-1 Random Folds}
\usage{
repeatedRandomKm1Folds(k, N)
}
\arguments{
\item{k}{number of folds}

\item{N}{population size}
}
\value{
A list of length \code{k} where each element is a vector
containing the indices of the sampled data.
}
\description{
Takes repeated samples from the population.
See the \strong{Details} section below for further information.
}
\details{
Each element in the population is randomly assigned to one of the k
folds by using the \code{\link{randomKFolds}} function.
A list of length k is then created from these folds, so that the i-th
item of the list is a vector of indices generated by removing the i-th fold
and merging the remaining k - 1 folds together.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Assign data to 3 folds
repeatedRandomKm1Folds(N = 10, k = 3)

}
\seealso{
\code{\link{randomKFolds}}
}
\author{
Alessandro Barberis
}
