% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedSample}
\alias{repeatedSample}
\title{Repeated Samples}
\usage{
repeatedSample(k = 1, sampleFun, ...)
}
\arguments{
\item{k}{integer, the number of repeated samples to generate}

\item{sampleFun}{the sampling function}

\item{...}{further arguments to \code{sampleFun}}
}
\value{
A list of length \code{k} where each element
is a vector of indices representing the sampled data.
}
\description{
Takes repeated samples from the population by using the provided
sampling function.
}
\examples{
\dontrun{
#set seed for reproducibility
set.seed(seed = 5381L)

#Take two samples without replacement
repeatedSample(
 k = 2,
 sampleFun = sampleWithoutReplacement,
 N = 10,
 n = 5
)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
