% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedSimpleRandomSampleWithoutReplacement}
\alias{repeatedSimpleRandomSampleWithoutReplacement}
\title{Repeated Random Samples}
\usage{
repeatedSimpleRandomSampleWithoutReplacement(k, N, n)
}
\arguments{
\item{k}{integer, the number of repeated samples to generate}

\item{N}{positive integer value, the population size}

\item{n}{positive integer value, the sample size}
}
\value{
A list of length \code{k} where each element
is a vector of indices representing the sampled data.
}
\description{
Takes repeated samples without replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
This function takes repeated samples by using the \code{\link{replicate}}
function and the \code{\link{simpleRandomSampleWithoutReplacement}} sampling
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Take two samples without replacement
repeatedSimpleRandomSampleWithoutReplacement(
 k = 2,
 N = 10,
 n = 6
)

}
\seealso{
\code{\link{simpleRandomSampleWithoutReplacement}}
}
\author{
Alessandro Barberis
}
