% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedStratifiedKm1Folds}
\alias{repeatedStratifiedKm1Folds}
\title{Repeated k-1 Stratified Folds}
\usage{
repeatedStratifiedKm1Folds(k, strata)
}
\arguments{
\item{k}{number of folds}

\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}
}
\value{
A list of length \code{k} where each element is a vector
containing the indices of the sampled data.
}
\description{
Takes repeated stratified samples from the population.
See the \strong{Details} section below for further information.
}
\details{
Each element in the population is randomly assigned to one of the k
folds - so that the percentage of each stratum in the population is preserved
in each fold - by using the \code{\link{stratifiedKFolds}} function.
A list of length k is then created from these folds, so that the i-th
item of the list is a vector of indices generated by removing the i-th fold
and merging the remaining k - 1 folds together.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(1,1,1,2,2,2,2,2,2)

#Check ratio
table(strata)/length(strata)

#Assign data to 3 folds
i = repeatedStratifiedKm1Folds(
 strata = strata,
 k = 3
)
#Check indices
i
#Check ratio in the samples made of k-1 folds
table(strata[i[[1]]])/length(strata[i[[1]]])
table(strata[i[[2]]])/length(strata[i[[2]]])
table(strata[i[[3]]])/length(strata[i[[3]]])

}
\seealso{
\code{\link{stratifiedKFolds}}
}
\author{
Alessandro Barberis
}
