% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-resampling-functions.R
\name{repeatedStratifiedSampleWithReplacement}
\alias{repeatedStratifiedSampleWithReplacement}
\title{Repeated Stratified Random Samples With Replacement}
\usage{
repeatedStratifiedSampleWithReplacement(k, strata, n, prob = NULL)
}
\arguments{
\item{k}{integer, the number of repeated samples to generate}

\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{n}{positive integer value, the sample size}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}
}
\value{
A list of length \code{k} where each element
is a vector of indices representing the sampled data.
}
\description{
Takes repeated stratified samples with replacement from the
population.
See the \strong{Details} section below for further information.
}
\details{
This function takes repeated samples by using the \code{\link{replicate}}
function and the \code{\link{stratifiedSampleWithReplacement}} sampling
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(rep("a", 3),rep("b", 6))

#Check ratio
table(strata)/length(strata)

#Stratified random samples
i = repeatedStratifiedSampleWithReplacement(
  k = 2,
  strata = strata,
  n = 3,
)
#Check ratio
table(strata[i[[1]]])/length(strata[i[[1]]])
table(strata[i[[2]]])/length(strata[i[[2]]])

}
\seealso{
\code{\link{stratifiedSampleWithReplacement}}
}
\author{
Alessandro Barberis
}
