% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-resampling-class.R
\name{resampling}
\alias{resampling}
\title{Constructor of the \code{resampling} class}
\usage{
resampling(
  method = c("rswor", "srswor", "stratified_rswor", "balanced_rswor", "permutation",
    "kfolds", "stratified_kfolds", "balanced_kfolds", "leave_p_out", "leave_one_out",
    "rswr", "srswr", "stratified_rswr", "balanced_rswr", "bootstrap"),
  N = integer(),
  removed = integer(),
  samples = list(integer())
)
}
\arguments{
\item{method}{character, the id of the used sampling method}

\item{N}{integer, the population size}

\item{removed}{integer, the indices of elements removed prior to sampling}

\item{samples}{list, each element is an integer vector containing the samples}
}
\value{
An object of class \code{resampling}.
}
\description{
This is the constructor of the \code{resampling} class.
}
\details{
An object of class \code{resampling} is a named list with 4 elements:
\describe{
\item{\code{method}}{the id of the used sampling method}
\item{\code{N}}{the size of the population from which the samples were taken.
Elements in the population have index from 1 to N}
\item{\code{removed}}{(optional) vector of indices of elements removed from the
population before taking the samples}
\item{\code{samples}}{list of samples repeatedly taken from the population. Each
element of the list is an integer vector containing the indices of the
elements sampled from the population}
}

Functions to facilitate access to the data stored in a \code{resampling} object are
available:
\itemize{
\item \code{?getSamplingMethodId}: returns the sampling method id
\item \code{?getPopulationSize}: returns the population size
\item \code{?getRemovedElements}: returns the indices of elements removed from the population
\item \code{?getNumberOfSamples}: returns the number of taken samples
\item \code{?getSamples}: returns the list of samples
\item \code{?getSampleSize}: returns a vector of integer values, the size of each sample
\item \code{?getHoldOutSample}: returns a list where each item is an integer vector
containing the indices of the elements not sampled from the population
\item \code{?getHoldOutSampleSize}:returns a vector of integer values, the size of each
hold-out sample
}

Other useful functions include:
\itemize{
\item \code{\link[=print.resampling]{print}}: print a summary of the \code{resampling} object
\item \code{\link[=plot.resampling]{plot}}: plot the samples taken from the population as an heatmap
}
}
\examples{
#default
resampling()

#resampling object
resampling(
 method = 'rswor',
 N = 10,
 samples = list(c(1,5,7))
)

}
\author{
Alessandro Barberis
}
