% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{sampleWithoutReplacement}
\alias{sampleWithoutReplacement}
\title{Random Samples}
\usage{
sampleWithoutReplacement(N, n, prob = NULL)
}
\arguments{
\item{N}{positive integer value, the population size}

\item{n}{positive integer value, the sample size}

\item{prob}{(optional) N-length vector of positive numeric values, the
probability weights for obtaining the \code{N} elements}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a sample without replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
This function is a wrapper to the \code{\link[base]{sample.int}}
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Simple random sample without replacement
sampleWithoutReplacement(N = 10, n = 5)

}
\author{
Alessandro Barberis
}
