% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{simpleRandomSampleWithoutReplacement}
\alias{simpleRandomSampleWithoutReplacement}
\title{Random Samples}
\usage{
simpleRandomSampleWithoutReplacement(N, n)
}
\arguments{
\item{N}{positive integer value, the population size}

\item{n}{positive integer value, the sample size}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a simple random sample without replacement from the population.
\emph{Simple random sampling} (\emph{SRS}) is the easiest form of sampling where each
element of the population has the same probability of being selected for the
sample.
See the \strong{Details} section below for further information.
}
\details{
This function is a wrapper to the \code{\link[base]{sample.int}}
function.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Simple random sample without replacement
simpleRandomSampleWithoutReplacement(N = 10, n = 5)

}
\author{
Alessandro Barberis
}
