% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{stratifiedKm1Folds}
\alias{stratifiedKm1Folds}
\title{K-1 Stratified Folds}
\usage{
stratifiedKm1Folds(k, strata, i = NULL)
}
\arguments{
\item{k}{number of folds}

\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{i}{(optional) integer, fold to be use as holdout data}
}
\value{
A vector containing the indices of the sampled data.
}
\description{
Assigns the population to k stratified folds and returns a random
sample made of elements from k-1 folds.
See the \strong{Details} section below for further information.
}
\details{
Each element in the population is randomly assigned to one of the k
folds - so that the percentage of each stratum in the population is preserved
in each fold - by using the \code{\link{stratifiedKFolds}} function.
If provided, \code{i} indicates the i-th fold to be considered as holdout data.
If \code{i} is missing, one fold is randomly selected to be the holdout data.
A random sample is then generated by removing the i-th fold and merging the
remaining k - 1 folds together.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(1,1,1,2,2,2,2,2,2)

#Check ratio
table(strata)/length(strata)

#Assign data to 3 folds
i = repeatedStratifiedKm1Folds(
 strata = strata,
 k = 3
)
#Check indices
i
#Check ratio in the samples made of k-1 folds
table(strata[i[[1]]])/length(strata[i[[1]]])
table(strata[i[[2]]])/length(strata[i[[2]]])
table(strata[i[[3]]])/length(strata[i[[3]]])

}
\seealso{
\code{\link{stratifiedKFolds}}
}
\author{
Alessandro Barberis
}
