% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sampling-functions.R
\name{stratifiedSampleWithoutReplacement}
\alias{stratifiedSampleWithoutReplacement}
\title{Stratified Random Samples}
\usage{
stratifiedSampleWithoutReplacement(strata, n, prob = NULL)
}
\arguments{
\item{strata}{vector of stratification variables. The population size is
\code{length(strata)}}

\item{n}{positive integer value, the sample size}

\item{prob}{(optional) vector of positive numeric values, the probability weights
for obtaining the \code{strata} elements. If provided, it must be the same
length as \code{strata}}
}
\value{
A vector of length \code{n} containing the index of the
computed random set of observations.
}
\description{
Takes a stratified sample without replacement from the population.
See the \strong{Details} section below for further information.
}
\details{
Stratified sampling is a technique of sampling from a population that
can be partitioned into 'strata' (or 'subpopulations'), where each element in
the population is part of one and only one stratum. It is used to ensure that
subgroups of the population are represented in the taken sample.
This function implements the so-called "proportionate allocation", in which
the proportion of the strata in the population is maintained in the samples.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define strata
strata = c(rep("a", 3),rep("b", 6))

#Check ratio
table(strata)/length(strata)

#Stratified random sample
i = stratifiedSampleWithoutReplacement(
  strata = strata,
  n = 3
)
#Check indices
i
#Check ratio in the sample
table(strata[i])/length(strata[i])

}
\references{
https://en.wikipedia.org/wiki/Stratified_sampling
}
\author{
Alessandro Barberis
}
