% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrang.R
\name{subrang}
\alias{subrang}
\title{Second-step Resampling Function}
\usage{
subrang(SEED, N, n, j, sam, type)
}
\arguments{
\item{SEED}{a single value, interpreted as an integer.}

\item{N}{a positive integer giving the number of iterations.}

\item{n}{a non-negative integer giving the number of items to choose.}

\item{j}{a positive integer representing the row index of matrix sam to
choose from.}

\item{sam}{matrix attained by first-step resampling.}

\item{type}{character indicating the type of data, e.g., "luad".}
}
\value{
a matrix that each row represents the indices of samples for
        each iteration.
}
\description{
This function is used for constructing sample subsets from the
matrix of second-step resampling.
}
\examples{
subrang(SEED=1, N=5, n=3, j=2, sam=sam, type="sarc")
}
\keyword{resampling}
